+++ src/resources/extensions/gsd/migrate/parsers.ts
@@ -3,7 +3,7 @@
 // Zero Pi dependencies — uses only exported helpers from files.ts.
 
 import { splitFrontmatter, parseFrontmatterMap, extractBoldField } from '../files.js';
-import { normalizeStringArray } from '../../shared/index.js';
+import { normalizeStringArray } from '../../shared/mod.js';
 
 import type {
   PlanningRoadmap,
