+++ src/resources/extensions/gsd/queue-reorder-ui.ts
@@ -11,7 +11,7 @@
 import type { ExtensionContext } from "@gsd/pi-coding-agent";
 import { type Theme } from "@gsd/pi-coding-agent";
 import { Key, matchesKey, truncateToWidth, type TUI } from "@gsd/pi-tui";
-import { makeUI, GLYPH } from "../shared/index.js";
+import { makeUI, GLYPH } from "../shared/mod.js";
 import { validateQueueOrder, type DependencyValidation } from "./queue-order.js";
 
 export interface ReorderItem {
