+++ src/resources/extensions/gsd/triage-ui.ts
@@ -10,7 +10,7 @@
  */
 
 import type { ExtensionCommandContext } from "@gsd/pi-coding-agent";
-import { showNextAction } from "../shared/index.js";
+import { showNextAction } from "../shared/mod.js";
 import type { CaptureEntry, Classification, TriageResult } from "./captures.js";
 import { markCaptureResolved } from "./captures.js";
