+++ src/resources/extensions/shared/mod.ts
